requirements.txt
mysql-connector-python==8.0.33
configparser==5.3.0